requirements.txt
asgiref==3.8.1
asttokens==3.0.0
black==24.10.0
click==8.1.7
decorator==5.1.1
Django==4.2.17
django-cors-headers==4.6.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.1.0
ipdb==0.13.11
ipython==8.30.0
jedi==0.19.2
Markdown==3.7
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.10.1
pytz==2024.2
sqlparse==0.5.2
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
